api/requirements.txt
uvicorn
fastapi
Pillow
tqdm
scikit-learn
python-multipart
timm
